delete card router fix

+++ src/controllers/card.ts
@@ -15,22 +15,31 @@ export const getCards = async (_req: Request, res: Response, next: NextFunction)
   }
 };
 
-export const deleteCardById = async (req: Request | any, res: Response, next: NextFunction) => {
+export const deleteCardById = async (
+  req: Request | any,
+  res: Response,
+  next: NextFunction,
+) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findById(cardId);
-
-    // здесь извлекаю id из ObjectId
-    const isCardOwner = (card?.owner as any).toHexString() === req.user._id;
-    if (isCardOwner) {
-      const cardToDelete = await Card.findByIdAndDelete(cardId);
-      if (cardToDelete) {
-        res.status(REQUEST_STATUS.OK).send(card);
+    if (card) {
+      // здесь извлекаю id из ObjectId
+      const isCardOwner = (card?.owner as any).toHexString() === req.user._id;
+      // проверяю владельца
+      if (isCardOwner) {
+        const deletedCard = await Card.findByIdAndDelete(cardId);
+        // финально проверяю что карточка удалена
+        if (!deletedCard) {
+          next(new NotFoundError('Неизвестная ошибка удаления карточки'));
+        } else {
+          res.status(REQUEST_STATUS.OK).send(card);
+        }
       } else {
-        next(new NotFoundError('Не найдет удаляемый документ'));
+        next(new NotAuthorizedError('Нет прав для удалений такой карточки'));
       }
     } else {
-      next(new NotAuthorizedError('Нет прав для удалений такой карточки'));
+      next(new NotFoundError('Не найден удаляемый документ'));
     }
   } catch (error) {
     if (error instanceof MongooseError.CastError) {
